% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumber.R
\name{renumber}
\alias{renumber}
\title{renumbers files and directories}
\usage{
renumber(
  path = ".",
  number_dir = TRUE,
  prefix = "",
  postfix = "-",
  old_prefix = "",
  test_only = FALSE
)
}
\arguments{
\item{path}{path to renumber. Defaults to working directory.}

\item{number_dir}{Logical. If TRUE, directories will be renumbered, otherwise files.}

\item{prefix}{Anything to go before the number. Defaults to an empty string.}

\item{postfix}{What to go after the number. Defaults to a hyphen}

\item{old_prefix}{Any previous numbering system that should be renumbered.}

\item{test_only}{If TRUE will output data.frame showing changes that would be made. Files will be untouched.}
}
\description{
Numbered files help the user workout the order in which files should be used. This function makes it easy to number or renumber files with a gadget.
}
\examples{
\donttest{
renumber_gadget(path = ".", number_dir = FALSE, test_only = TRUE)
}
}
